requirements.txt
Flask
SQLAlchemy
torch
torchvision
Pillow
numpy
scikit-learn
Werkzeug

